refactor(package): move frameScore to frame.score
issue [#152(comment)](https://github.com/next-step/java-bowling/pull/151#discussion_r4177739030
