+++ src/main/java/bowling/domain/Frames.java
@@ -2,7 +2,7 @@ package bowling.domain;
 
 import bowling.domain.frame.Frame;
 import bowling.domain.frame.NormalFrame;
-import bowling.domain.frameScore.FrameScore;
+import bowling.domain.frame.score.FrameScore;
 
 import java.util.ArrayList;
 import java.util.Collection;
