+++ src/main/java/bowling/domain/frame/Frame.java
@@ -1,7 +1,7 @@
 package bowling.domain.frame;
 
 import bowling.domain.ShotScore;
-import bowling.domain.frameScore.FrameScore;
+import bowling.domain.frame.score.FrameScore;
 
 import java.util.List;
