+++ src/main/java/bowling/domain/frame/score/DefaultFrameScore.java
@@ -1,4 +1,4 @@
-package bowling.domain.frameScore;
+package bowling.domain.frame.score;
 
 import bowling.domain.scoreType.ScoreType;
