+++ src/test/java/bowling/domain/frame/FinalFrameTest.java
@@ -1,6 +1,5 @@
 package bowling.domain.frame;
 
-import bowling.domain.frame.FinalFrame;
 import bowling.domain.scoreType.ScoreType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
